@@ -15,18 +15,18 @@
   "type": "module",
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.3.11",
+    "@angular/animations": "14.0.2",
     "@angular/cdk": "13.3.9",
-    "@angular/common": "13.3.11",
-    "@angular/compiler": "13.3.11",
-    "@angular/core": "13.3.11",
+    "@angular/common": "14.0.2",
+    "@angular/compiler": "14.0.2",
+    "@angular/core": "14.0.2",
     "@angular/fire": "^7.4.1",
     "@angular/flex-layout": "12.0.0-beta.35",
-    "@angular/forms": "13.3.11",
+    "@angular/forms": "14.0.2",
     "@angular/material": "13.3.9",
-    "@angular/platform-browser": "13.3.11",
-    "@angular/platform-browser-dynamic": "13.3.11",
-    "@angular/router": "13.3.11",
+    "@angular/platform-browser": "14.0.2",
+    "@angular/platform-browser-dynamic": "14.0.2",
+    "@angular/router": "14.0.2",
     "core-js": "^3.8.1",
     "firebase": "^8.2.1",
     "rxjs": "~6.6.3",
@@ -34,10 +34,10 @@
     "zone.js": "~0.11.6"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.3.8",
-    "@angular/cli": "13.3.8",
-    "@angular/compiler-cli": "13.3.11",
-    "@angular/language-service": "13.3.11",
+    "@angular-devkit/build-angular": "14.0.2",
+    "@angular/cli": "14.0.2",
+    "@angular/compiler-cli": "14.0.2",
+    "@angular/language-service": "14.0.2",
     "@angular/localize": "^v12-lts",
     "@types/jasmine": "~3.6.2",
     "@types/jasminewd2": "~2.0.8",
